cbt_dbt_rag_assistant/pyproject.toml
[project]
name = "cbt-dbt-rag-assistant"
version = "0.1.0"
description = "CBT/DBT RAG Assistant"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pydantic (>=2.11.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pgvector (>=0.4.0,<0.5.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "typer (>=0.15.2,<0.16.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "streamlit (>=1.44.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
